Perf

before
```
sg scan -c ./sgconfig.yml ../TypeScript/src  17.63s user 0.46s system 167% cpu 10.823 total
```

after
```
sg scan -c ./sgconfig.yml ../TypeScript/src  9.95s user 0.38s system 432% cpu 2.387 total
```